pyproject.toml
[project]
name = "ocr-chess"
version = "0.1.0"
description = "Computer Vision Chess Assistant"
authors = [
    { name = "Dan Parizher", email = "dparizher@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">3.13"
dependencies = [
    "numpy",
    "opencv-python",
    "pillow",
    "pyqt6",
    "python-chess",
    "torch",
    "torchvision",
    "xxhash",
]

[project.optional-dependencies]
dev = [
    "deptry",
    "ruff",
    "uv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.deptry]
package_module_name_map = { python-chess = ["chess"] }
